@@ -126,11 +126,11 @@ namespace Test78
     public async Task TestLogstashServerLog78()
     {
       // ����
-      var logstashUrl = "http://192.168.32.122:5000";
+      var logstashUrl = "http://192.168.31.122:5000";
       var logstashLogger = new LogstashServerLog78(logstashUrl);
       var log = Log78.Instance;
       log.setup(logstashLogger, null, null);
-      log.LevelApi = 0; // ȷ��������־���ᱻ���͵� Logstash
+      log.LevelApi = 50; // ȷ��������־���ᱻ���͵� Logstash
 
       // ����������־��Ŀ
       var testEntry = new LogEntry
